// src/api-utils/generated-client/api.ts
export interface BankAccount {
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'uuid': string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'finicity_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'account_last_four'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'bank_account_number'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'bank_routing_number'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'bank_account_type'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'bank_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'account_salesforce_id'?: string;
    /**
     * 
     * @type {number}
     * @memberof BankAccount
     */
    'finicity_account_number_last_four'?: number;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'plaid_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'institution_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof BankAccount
     */
    'logo'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BankAccount
     */
    'is_real_bank_numbers'?: boolean;
}
// ...
export interface UpdateBankAccountDto {
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'uuid'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'finicity_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'account_last_four'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'bank_account_number'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'bank_routing_number'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'bank_account_type'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'bank_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'account_salesforce_id'?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateBankAccountDto
     */
    'finicity_account_number_last_four'?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'plaid_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'institution_name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateBankAccountDto
     */
    'logo'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof UpdateBankAccountDto
     */
    'is_real_bank_numbers'?: boolean;
}